feat: dockerize crawler
