+++ docker-house.yml
@@ -1,0 +1,58 @@
+version: '2.2'
+
+services:
+  selenium-hub:
+    image: selenium/hub:3.12.0-boron
+    container_name: selenium-hub
+    ports:
+      - "4444:4444"
+    networks:
+      - crawler
+
+  chrome:
+    image: selenium/node-chrome:3.12.0-boron
+    depends_on:
+      - selenium-hub
+    environment:
+      - HUB_HOST=selenium-hub
+      - HUB_PORT=4444
+    networks:
+      - crawler
+
+  crawler_taipei:
+    build:
+      context: ./
+      dockerfile: Dockerfile.crawler
+    container_name: crawler_taipei
+    environment:
+      - CITY_ID=1
+      - PAGE_LIMIT=300
+      - URL_END=9000
+      - EXECUTED_TIME=19:51
+      - CHROME_HOST=selenium-hub
+      - USE_DOCKER=1
+    networks:
+      - crawler
+
+    depends_on: [selenium-hub, chrome]
+
+  crawler_new_taipei:
+    build:
+      context: ./
+      dockerfile: Dockerfile.crawler
+    container_name: crawler_new_taipei
+    environment:
+      - CITY_ID=3
+      - PAGE_LIMIT=300
+      - URL_END=9000
+      - EXECUTED_TIME=19:51
+      - CHROME_HOST=selenium-hub
+      - USE_DOCKER=1
+    networks:
+      - crawler
+
+    depends_on: [selenium-hub, chrome]
+
+networks:
+  crawler:
+    driver: bridge
